+++ data-transfer/src/test/java/it/eng/datatransfer/rest/protocol/ConsumerDataTransferCallbackControllerTest.java
@@ -21,6 +21,7 @@ import it.eng.datatransfer.exceptions.TransferProcessNotFoundException;
 import it.eng.datatransfer.model.Serializer;
 import it.eng.datatransfer.model.TransferCompletionMessage;
 import it.eng.datatransfer.model.TransferStartMessage;
+import it.eng.datatransfer.model.TransferSuspensionMessage;
 import it.eng.datatransfer.service.DataTransferService;
 import it.eng.datatransfer.util.MockObjectUtil;
 import jakarta.validation.ValidationException;
@@ -90,6 +91,7 @@ public class ConsumerDataTransferCallbackControllerTest {
 				Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_COMPLETION_MESSAGE)));
 	}
 	
+	// terminate transfer
 	@Test
 	@DisplayName("Terminate TransferProcess")
 	public void terminateDataTransfer() {
@@ -106,19 +108,32 @@ public class ConsumerDataTransferCallbackControllerTest {
 					Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_START_MESSAGE)));
 	}
 	
+	// suspend transfer
 	@Test
 	@DisplayName("Suspend/pause TransferProcess")
 	public void suspenseDataTransfer() {
-		assertEquals(HttpStatus.NOT_IMPLEMENTED, 
-				controller.suspenseDataTransfer(MockObjectUtil.CONSUMER_PID,
-						Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_SUSPENSION_MESSAGE)).getStatusCode());
+		when(dataTransferService.suspendDataTransfer(any(TransferSuspensionMessage.class), any(String.class), isNull()))
+		.thenReturn(MockObjectUtil.TRANSFER_PROCESS_COMPLETED);
+		ResponseEntity<JsonNode> response = controller.suspenseDataTransfer(MockObjectUtil.CONSUMER_PID,
+				Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_SUSPENSION_MESSAGE));
+		assertEquals(HttpStatus.OK, response.getStatusCode());	
 	}
 	
 	@Test
-	@DisplayName("Terminate TransferProcess - invalid request body")
+	@DisplayName("Suspend TransferProcess - invalid request body")
 	public void suspenseDataTransfer_invalidBody() {
 		assertThrows(ValidationException.class, () ->
 			controller.suspenseDataTransfer(MockObjectUtil.CONSUMER_PID, 
 					Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_START_MESSAGE)));
+	}
+	
+	@Test
+	@DisplayName("Suspend TransferProcess - error service")
+	public void suspendDataTransfer_errorService() {
+		when(dataTransferService.suspendDataTransfer(any(TransferSuspensionMessage.class), any(String.class), isNull()))
+			.thenThrow(TransferProcessNotFoundException.class);
+		assertThrows(TransferProcessNotFoundException.class, 
+				() -> controller.suspenseDataTransfer(MockObjectUtil.PROVIDER_PID,
+				Serializer.serializeProtocolJsonNode(MockObjectUtil.TRANSFER_SUSPENSION_MESSAGE)));
 	}
 }